Support replicas in container runtime jobs.
